ds2: Delete cache for product when starting a new build instruction
If we don't do that, we might continue linking to documents even if they
do not build anymore and thus will not be available.

I guess the negative consequence of this is that it will be less likely
now that you'll get the document titles into the status JSON file.