build-nav: Build navigation only for documents that have been built
We previously would build navigation entries even for docsets that had
never been built, leading to the issue that depending on build state,
we might show a lot pages that were empty apart from the product
description.

In addition, this solution had the issue that it could soft-unpublish
documents that were not mentioned in the product config anymore but
whose docsets had not yet been rebuilt.

We now also exclude documents based on their @lifecycle value and
whether @navigation-visible=hidden is set