stitch-schema: Remove <styleroot/> element
<styleroot/> was only supported in the schema but never in the docserv2
itself or in daps2docker. There also was no use case for it thus far.
(Granted, if we ever needed to publish the DAPS documentation using
Docserv2, that might be a use case. Though we could also just add a new
set of DC files that has the DAPS-brand styleroot.)