d2d: Make Docserv2 actually use DAPS parameters like --draft
We would previosuly generate these parameter but not properly write
them to a file. I think the fault for that lies in how the subprocess
command in Deliverable.execute is configured, with it piping command
line output to itself instead of allowing redirection into a file.
That is only a hunch though.