stitch-schema: Remove "default" attribute from docset
The default attribute of docsets has not proven to be helpful: We need
to create (additional) redirects and configuration for Apache manually
in any case and the .htaccess file format does not support including
additional files.

Frank suggested to use the default attribute product version sorting on
the index page. However, I don't really see the use case for that
anymore now that there is the sort-priority attribute.