ds2: Generate names of local repo dirs based on MD5 hash
The previous algorithm that just replaced a number of special characters
in Git remote URLs with underscores could have easily generated colliding
repo name dirs. The list of special characters that it replaced was not
particularly exhaustive either.
The new version just runs the remote URL through the MD5 algorithm and
then attaches the last part of the URL, so the repo can easily be
identified by humans too.