deliverable.py: Correctly export titles with &, <, >
The titles we get from xmlstarlet contain & < > as entities (curiously,
the same is not true of " and '). The cElementTree function that later
writes out our XML cache files then encodes those entities again, as
e.g. as &amp;amp;.
Hence, let's try to work around that by replacing the entities we get
from xmlstarlet with the literal character.