schema: Allow specific deliverables to come from different subdirs
You can now configure stacked subdirs for a specific language and its
individual deliverables. E.g.:

If <language> includes <subdir>l10n/ja-jp</subdir> and
<deliverable> includes <subdir>quickstart</subdir>
  -> We check the directory `l10n/ja-jp/quickstart`

If there is a <subdir/> element in both the deliverable of the default
language and the equivalent deliverable of a translation language, the
latter deliverable's <subdir/> will overwrite the former's.
However, where it is not overwritten, the <subdir/> value is
inheritable from the default language's deliverable.