schema/validation/bih: docset/@navigation can disable navigation entirely
This is yet another rename of the @navigation-visible attribute and it
now allows for completely disabling navigation builds via
docset/@navigation="linked|hidden|disabled".

* "linked" - previously called "visible"; the default, navigation for
  the docset is generated and linked to from the homepage
* "hidden" - previously also called "hidden"; navigation for the docset
  is built, but not linked to from the homepage
* "disabled" - new!; navigation is not built or linked

This change also removes the product/@navigation-visible attribute from
the schema, as it has never actually been supported in the code.
(Enabling it via simplify.xsl would not be hard, but then again, it
apparently has never been needed so far.)