bih/deliv/build-nav: Allow running BIs when there's no <builddocs> (#131)
In the past, Docserv2 would not accept build instructions for products
like SUMA that only have <internal> <ref>s and <external> <link>s in
their docsets, because they come prebuilt.

The solution to that was that we always rebuilt the SUMA pages,
whenever any other navigation page was being rebuilt, which led to a
quite a few extra rebuilds of the VMDP docs (which are small and don't
change much in terms of content).