Support "gated" attribute in product config schema
The Docserv config allows the "gated" attribute (type boolean)
on these elements:

* <product>
* <docset>
* <deliverable>
* <link>

By default, it's set to gated="false" (= publicly available).

The idea to have it on these different levels is flexibility and readability.
You can have a product and set everything as gated="true", but still have a
few deliverables that are publicly available.