Add TODO and prepare for Python stitching
The Bash implementation of bin/docserv-stitch is too slow.
We need to pass it to the lxml library (written in C!) to have
a more efficient way of doing this.

Maybe it's even better to stitch first and do all sorts of validation
(references, unique productid etc.) afterwards.
Could it be a memory issue?